Split out admin.php into latest- and v5-down versions; updated get_relevant_version() to accommodate admin requests
